part2/backend/requirements.txt
flask
pymongo
python-dotenv
flask-cors
flask-caching
redis
gunicorn
